chore(daemon): rename gateway to daemon
