+++ app/daemon/src/lib.rs
@@ -2,12 +2,9 @@
 //! scheduling. Personal agent, local-first.
 
 pub mod config;
-pub mod gateway;
+pub mod daemon;
 pub(crate) mod hook;
 
 pub use config::DaemonConfig;
-pub use gateway::{
-    Gateway,
-    handle::{ServeHandle, serve, serve_with_config},
-};
-pub use hook::GatewayHook;
+pub use daemon::{Daemon, DaemonHandle};
+pub use hook::DaemonHook;
